move all controller as rest controller
