+++ src/main/java/itu/jca/eval/test/coworking/controller/rest/OptionController.java
@@ -1,0 +1,8 @@
+package itu.jca.eval.test.coworking.controller.rest;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class OptionController {
+    
+}
